Changes location of FutureResult & ReposRequest on their own files
