+++ DataAccess/Public/CMakeLists.txt
@@ -9,7 +9,9 @@ set(SOURCE_FILES
         Include/DataAccess/IRepository.hpp
         Include/DataAccess/IServerRepositories.hpp
         Include/DataAccess/SQLStatements.hpp
-        Include/DataAccess/QueryBuilder.hpp)
+        Include/DataAccess/QueryBuilder.hpp
+        Include/DataAccess/RepositoryRequest.hpp
+        Include/DataAccess/FutureResult.hpp)
 
 add_library(${TARGET} INTERFACE)
